refactor(library/Attempt): moves "Fresh/synchronous.ts" to "Fresh/that.ts"
